NOTE: This is not production ready. It has all the major work done, but still needs a lot of testing. v1.0.3 should still be used in production.
Pretty major commit, this one is. All the SSH plumbing has been replaced so that passwordless SSH is not longer used. The funtions that handled adding the dashboard's apapche user's keys are gone.

Added a new option to each cluster definition to allow for users to set a separate root password from their ricci password, if needed. If the root password is not set, the password set for ricci is copied.

The new Net::SSH2 perl module has been compiled as an RPM and added to the new AN!Repo. Anyone updating will need to add it.

Added a new CHANGELOG file as well.

This marks a feature freeze for v1.0.4.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>