NOTE: This is not production ready. It has all the major work done, but still needs a lot of testing. v1.0.3 should still be used in production.
This should be the last commit with regard to the SSH plumbing changes, save for debugging. Connections will now stay open for as long as the remote system holds the connection open. Closing on either STDOUT or STDERR are also now caught.

Special thanks to Rafael Kitover (rkitover@gmail.com), maintainer of Net::SSH2, for helping me sort out the polling and data collection.

With luck, the next commit will be for the 1.0.4 release.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>