TE: This is not production ready. It has all the major work done, but still needs a lot of testing. v1.0.3 should still be used in production.
Ok, so about being close to done...

This commit fixes several bugs, but more directly, it now starts a fairly major update to the new configuration section.

Instead of showing the config options for all Anvil!'s on one page, a list of Anvil!s are now shown and the user clicks to modify an existing one or add a new one.

This was done so that the same "global" options can be displayed under each Anvil! configuration section. This will soon allow users to set global values, like email recipients, and selectively override those values to given systems.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>