Note: This is a beta release. Please do not use in production.
This fixes a bug with saving the global options. It fixes a bug with how hosts is written. Renamed 'Media Connector' to 'Media Library'.

This also adds a new 'common.lib' library and new logos. This is the start of adding multi-lingual support in the next y-stream release. It has no impact on the current version.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>