NOTE: This commit is almost certainly broken, do not use outside of testing!
Oh boy, this is a big one!

First, this is laying the ground work for a renaming of the dashboard from 'AN!CDB' to 'Striker'.

This commit pays back some significant technical debt...

All text shown to the user has been moved to an XML file. A new function called 'get_string()' has been added that will pull out the string of text based on the current user's selected language. At this time, that's the system language, but at least now each install can be a given language and the dashboard is now relatively easily translated.

Skinning support has now been added via template files. How these are use will need to be documented which will be done later in a dedicated "Skinning" guide.

This commit is the first with all of the strings and website output converted to the new system. It is untested, outside of loading the main page without compilation errors. I will be testing and fixing over the next couple of days.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>