THIS COMMIT IS BROKEN!
Ok, so not so much broken as only half-done, but I haven't checked in for quite some time.

This check-in covers all <a ...> and <img ...> based buttons being now skinable. It also has the start of the media library becoming translatable and skinnable. It's not done. No sir, not at all.

Also, there is a known issue with guacamole 0.9.2 (https://glyptodon.org/jira/browse/GUAC-789) that seems to effect us. So that's broken, too.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>