Untested but should be working commit
Finished adding id="..." to links, images and input fields. Redid the node control panel to be a single template instead of multiple small ones. Will be doing the same to the other main page panels shortly.

Oh, guacamole is still broked, but that is making progress, too. Browsers are blocking things because of the 80 -> 8080 linking. Will try to resolve this with apache's mod_proxy.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>