Woot! The network configuration works (again)! This time, with more prompting to confirm goodness!
I also changed the anvil-map-network script to create backups when run, even in summary mode. I also added /etc/hosts and /etc/sysconfig/network to the files it backs up. This way, backups are made even if the MACs are already set in the install manifest and the user doesn't request a remap.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>