Major milestone hit!
Still very broken, but this commit actually partitioned the nodes. The sizing wasn't perfect and the second pool had to be auto-shrunk, but it's a good start.

Started work on the DRBD setup, but the next commit needs to handle a reboot after the partitioning and before the DRBD config.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>