I'm making a note here; Huge success!
Much more testing is still needed, but this commit includes a pretty major revamp of the partition creation logic and it seems to work very well in initial testing.

Nodes are properly, sequentially rebooted now if needed, once the stage1 storage (partitioning) is done. The sequential approach is used to help ensure remote access if one node fails to reboot. Safety over speed!

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>