This commit updates 'striker-installer'.
It adds two new features and fixes a few bugs.

* Striker can now setup PXE using a a disc in physical optical drive, simplifying setup for users who built their Striker dashboard from a phsycial install disc.
* Added a new function to 'freeze' network interfaces when the install starts.
* Changed the command line argument for RHN user and lease range to be "foo:bar" to match the rest of the switches.

The freeze function was needed for people running the Striker installer over the network on Striker dashboards running a GUI. NetworkManager would break the connection when the installer deleted the old files and wrote out the new ones. To get around this, the freeze function now looks for interfaces that are up and have an IP and immediately changes them to ONBOOT="yes" and NM_CONTROLLED="no", stops 'NetworkManager' and starts 'network'. Nothing else is changed at that point, so the network interfaces immediate go back to working without breaking SSH sessions.

There was a little more cleanup of what was printed to the screen, too.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>