This commit needs more testing.
This commit principally resolves around enabling Install Manifest to detect if the target node is not on the Internet and, if not, adds it's local repository/ies to nodes. It checks to see what it's BCN and IFN IPs are, what it's hostname is and whether it has a local generic, CentOS and/or RHEL repo. If a given node is off the Internet and if it is a repo host, it will add it's repos early in the install process. This should allow the install of packages needed to complete the setup of nodes when no Internet is available. Note that the local repository must contain all needed RPMs! generally, the "generic" repo is a clone of an-repo (https://alteeve.ca/an-repo/el6/).

Others:
* Fixed a bug where UPS IPs where set in the UPS name fields when setting up an install manifest.
* Dusted off the tools/gather-system-info tool and made it work on EL6.
* Converted the 'add_an_repo()' function to 'add_local_repo()' to handle the above-described tasks.
* Fixed a bug with adding NTP servers to nodes.
* Fixed a bug in tools/anvil-map-network where it would start to ask the user to unplug the cable they want to make..., but not print the name and go into a silent loop waiting for all interfaces to come up.
* Made striker-installer NOT overwrite existing PXE-based kickstart scripts, allowing users to customize their PXE kickstarts when the dashboard is itself installed from a kickstart script.
* Added 'skip_if_unavailable=1' to the repos we host. This way, if a node has repos for both dashboards, yum won't fail if one goes offline.
* Fixed an error in striker-installer's help page.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>