More testing needed, but this is a more stable check-in.
The major change in this check-in is that, now, newly provisioned servers are immediately added to the Anvil!. This required a pretty thorough revamp of how the server install messages/templates are handled. This opportunity was used to reduce the number of single-use template chunks. The installer auto-sets the HDD as the first boot device during the addition on the server to the Anvil!. This relieves the need to wait for the server to reboot for the first time.

Added two new variables for controlling what is showing during an Install Manifest run. You can now disable Internet checks and disable RHN registration. These are designed to avoid potentially worrying errors when an Anvil! is being provisioned in an environment where there will not be an Internet connection.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>