This commit adds the new 'anvil-kick-apc-ups' tool that can be used to turn one or more APC-brand UPSes with AP9630 network management cards as cluster-wide or rack-wide watchdog timers.
The idea is that, in some installations where physical access is severely restricted, this tool can be run every five minutes (or whatever) via cron. Each run, it will clear and previous power-off timers, set the selected sleep time and power off delays (5 minutes and 10 minutes, respectively), and initiate a timed power off. So long as at least one node is running, the system should never power off. However, if something blocks access (like a switch stack lock-up), the tool will no longer be able to "kick" the UPSes. After the poweroff delay expires, the UPS(es) will power off for the sleep time, then power back on.

When used along with ScanCore running 'nodemonitor' on the dashboards and with 'safe-anvil-start' on the nodes, the entire Anvil! (and all devices connected to the UPSes) will hard power cycle. The Anvil! will return to normal operation, with luck.

Carefully consider the risks of using this tool and ensure the problem it is trying to solve is a higher risk. This tool, should it fail, can and will cause a hard reset of the entire system!

Note: Once rebooted, the UPSes will not reset again until and unless this program restarts the process.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>