Big commit...
I'm starting to merge in Tom's Scanner code, but I wanted to use this as an opportunity to start paying off some serious technical debt. To do this, I've copied in the old AN::Tools project (it's in it's own repo, but I will probably not merge back for a while). This commit doesn't merge any of his code yet, but is mainly a new stand-alone "code" found in the "ScanCore/" directory. It pulls in the various modules from "AN/" and actually uses them for logging, DB connection, string parsing and so on.

The plan is, once ScanCore is merged/imported, to go back and start cleaning up Striker's existing code and ditch as much of it as possible in favour of the AN::Tools modules (moving in functions that might have general purpose). This will be a big project, and I may or may not release v2 before it is finished... We'll see. It's been a while and a new release is very much overdue.

Specific changes:
* New directories:
** ScanCore/ - See above
** AN/ - Perl modules; Imported from https://github.com/digimer/AN-Tools and heavily updated already.
** t/ - Currently very broken unit tests for the AN::Tools modules. Will eventually be the home of all Striker/ScanCore unit tests.
* Changed the Install Manifest run summary to use a new row name for PDUs (row_0284).
* Started work on making anvil-kick-apc-ups work on SURTD5000RMXLT3U
* Made the number of minutes between "kicks" in anvil-kick-apc-ups configurable.
* Added a function to anvil-kick-apc-ups that can conver ticks to seconds and back.
* Cleaned up the striker-installer help text some more.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>