New feature (in progress)
To enable nodes and dashboards being provisioned via Striker's "Install Target" function, it is helpful to give the new OS access to the internet to facilitate RHN registration and access to external repositories. To this end, there is a new striker-installer switch called '--router-mode'. When set, it will install and configure shorewall to provide masquarading (routing) from the BCN to the IFN (and out to the Internet).

As of this commit, shorewall install and configuration appears to be working, but more testing is needed. However, toggling the Install Target function in Striker will not (yet) control shorewall (and toggling iptables). This will be added in the next commit.

Other things:
* Disabled installing the Scanner RPM on Striker for now.
* Added a function to more sensibly determine the network, given an IP and it's subnet mask (copied from http://www.perlmonks.org/?node_id=465359)
* Added a check to see if NetworkManager exists before trying to chkconfig it off to avoid needless error messages.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>