Broken commit; changes to clustat are not properly registering yet.
As of now, scan-clustat has all the code needed to parse most all state changes and report them to the user in plain language. Example; "The server has migrated", "The server is booting" instead of "state change! X -> Y" as was the original plan. This will make the agents more tedious to write, but the generated alerts (email or local files for collection) far easier to understand by non-technical readers. As of now though, there is a bug where the values in the state variables are being mangled. That will be fixed in the next commit.

* Fixed a bug in AN::Tools::Alerts->register_alert(); where the message variables hash was being passed to the SQL statement instead of the generated variable string.
* Cleaned up some commenting to make log level 2 more readable.
* Fixed bad functions in the new ScanCore SQL tables.
* Added messages for the various possible service (VM and others) state changes.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>