In the ongoing attempt to migrate to DRBD 8.4, I hit a snag. I can't get dual-primary DRBD working with the 'drbd.sh' RA, and using drbd as a script service, the way the Anvil! always has, runs into a timing issue where it passed on to clvmd too fast. So clvmd starts before the DRBD resources are ready, fails to find it's PV and freaks out (see: https://bugzilla.redhat.com/show_bug.cgi?id=1157951#c9). To work around this, this commit introduces some quality duct tape with the 'wait-for-drbd' init.d pseudo-LSB script and 'anvil-wait-for-drbd'. The LSB script is now added to the storage service tree, in between drbd and clvmd. It's sole job is to sit there and not exit until all DRBD resources have gone dual-primary (with a timeout-and-error of 60 seconds).
In very initial testing, this appears to make things work better, but much more testing is needed.

There are changes to how shorewall configs are generated as well. Outside git, shorewall was upgraded to 4.6.11-1 in the AN!Repo and included a fix to shorewall's init.d script to fix 'status' always returning '0'. This appears to fix the problem with shorewall not starting properly when 'Install Target' is turned on when '--router-mode' was set during Striker's install.

Signed-off-by: Madison Kelly (digimer) <digimer@alteeve.ca>