IMPORTANT CHANGE! * The ScanCore loop sleep timer has been changed to 60 seconds and an example config has been added to striker.conf to simplify user overrides.
Major bug fix for a rare failure scenario; Grab a coffee if you want to read the detail;

In a configuration with four PDUs, two per rack, sourced from two UPSes where each rack has a PDU from either UPS, and if one of those PDUs fail, the attached node will lose half if it's power and also lose its backup fence method. Later, if the UPSes lose input power before the PDU is repaired, the UPS with two good UPSes will be under a much higher load than the other UPS (who has had about half of it's load transfered to the peer UPS). This will cause one UPS to deplete a lot faster than theother. Because the emergency shutdown is based on the assumption that both nodes have equal power, shut down will not occur because the half-loaded UPS will show lots of remaining power. When the weaker UPS depletes, however, one of the nodes will fail. At this time, the surviving peer has lost both fence method because IPMI was lost with the node and one of the PDUs is damaged/removed. At this point, ScanCore, via scan-clustat, will see the peer's state get set to 'Offline' before the fence succeeds. This caused the last-known state of the peer to be marked as "off". So later, if the second UPS depletes, the remaining node will fail. If the nodes have been configured to boot on AC restore, and if single-node-start is enabled, the second node to fail will boot long before the other node. It will see that the peer was last seen as off and try to run a single-node-start operation. If the second UPS charges too slowly, then anvil-safe-start's delays will expire, cman will start and try to fence the lost peer. IPMI will fail because the peer still has no AC and will move on to the PDU fence method. This will fail because the other PDU is still dead, and the "off" state on the good PDU will be retained.

* The fix, in scan-clustat, was to change the setting of each peer's last known state to be evaluated on every pass. If the peer's status is 'Offline, rgmanager' (which is the state before the fence is called), nothing will happen. If the state is just "Offline" though, the existing last known state is checked. If it is already 'off', nothing more happens. If it wasn't off though, a new method, 'Cman->check_for_pending_fence()' is called, which is turn uses 'fence_tool ls' to see if a fence is pending. If so, it informs the caller and nothing further happens. Only when there is no pending fence *and* the state is "Offline" will the last known state be set to 'off'. With this change, in the scenario above, the first node up will not attempt a single-node-start.

Other fixes and changes;
* Fixed a bug in 'Get->server_uuid()' where if a node was not in the Anvil! and it was selected to get the server's UUID, it would fail to get the server's XML and give up. Now, if no XML is returned from a node, it will query the database for the XML.
* Added a new '--clear-resync-flag' to ScanCore. This will allow the user to clear the DB resync flag if another instance died while the flag was set.
* Fixed a bug in ScanCore's update_anvil_data() function where reading in old servers' 'server_start_after' and 'server_start_delay' columns are set to NULL if nothing is returned in a given record's column.

Signed-off-by: Digimer <digimer@alteeve.ca>